Edit functions and add equals button functions
The add, subtract, divide, and multiply functions a part of the operate function could benefit from having the return tied to a "result" variable. Also combined the functions to get the first and second numbers for an operation, so as to allow for the first number to be inputted and then the second number based on certain conditions.

Also added functionality to the equals button to allow the operate function to execute on button click.